give classes own file
