+++ Models/Comments.cs
@@ -1,0 +1,7 @@
+namespace project_backend.Models;
+
+public class Comments
+{
+    public string User { get; set; }
+    public string Reply { get; set; }
+}
