+++ Models/Workout.cs
@@ -1,0 +1,7 @@
+namespace project_backend.Models;
+
+public class Workout
+{
+    public string? Title { get; set; }
+    public Lifts[]? Lifts { get; set; }
+}
